chore: move test files to dedicated folder
