--- lib/server/runner/index.ts
@@ -1,3 +1,0 @@
-import Runner from "./runner";
-
-export default Runner;
