+++ pages/api/team.ts
@@ -1,5 +1,5 @@
 import { Authentication } from "@lib/server/auth";
-import { MethodNotImplementedError, UnauthorizedError, UnsupportedMethodError } from "@lib/server/errors";
+import { UnauthorizedError, UnsupportedMethodError } from "@lib/server/errors";
 import { middleware } from "@lib/server/middleware";
 import prisma from "@lib/server/prisma";
 import { Role } from "@prisma/client";
