+++ tests/runner/data/fail.syntax-error.c
